Refactoring: create new folder for refactor, moved everything to it and changed references
